+++ src/main/java/racinggame/model/round/Round.java
@@ -1,0 +1,4 @@
+package racinggame.model.round;
+
+public class Round {
+}
