+++ src/test/java/racinggame/controller/InputValidatorTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Assertions;
 import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.model.validator.InputValidator;
 
 public class InputValidatorTest {
