+++ src/test/java/racinggame/controller/RoundValidatorTest.java
@@ -3,6 +3,7 @@ package racinggame.controller;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.model.validator.RoundValidator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
